# Conflicts:
#	src/backend/ci/ext/tencent/remotedev/biz-remotedev-tencent/src/main/kotlin/com/tencent/devops/remotedev/common/exception/ErrorCodeEnum.kt
#	src/backend/ci/ext/tencent/remotedev/biz-remotedev-tencent/src/main/kotlin/com/tencent/devops/remotedev/service/gitproxy/GitProxyTGitService.kt
#	support-files/i18n/remotedev/message_en_US.properties
#	support-files/i18n/remotedev/message_zh_CN.properties

+++ src/backend/ci/ext/tencent/environment/api-environment-tencent/src/main/kotlin/com/tencent/devops/environment/pojo/job/jobresp/StepHostResult.kt
@@ -31,5 +31,5 @@ data class StepHostResult(
     @get:Schema(title = "执行结束时间", description = "Unix时间戳，单位毫秒", required = true)
     val endTime: Long,
     @get:Schema(title = "总耗时", description = "单位毫秒", required = true)
-    val totalTime: Int
+    val totalTime: Long
 )
