+++ src/backend/ci/ext/tencent/environment/biz-environment-tencent/src/main/kotlin/com/tencent/devops/environment/pojo/job/jobcloudres/JobCloudStepHostResult.kt
@@ -41,5 +41,5 @@ data class JobCloudStepHostResult(
     val endTime: Long,
     @get:Schema(title = "总耗时", description = "单位毫秒", required = true)
     @JsonProperty("total_time")
-    val totalTime: Int
+    val totalTime: Long
 )
